+++ NEWS.md
@@ -1,3 +1,5 @@
+# labelled (development version)
+
 # labelled 2.10.0
 
 * `look_for()` now accepts `survey` objects (#121)